LayoutTests/fast/dom/shadow/adopt-node-with-shadow-root-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.adoptNode(shadow) threw exception Error: HierarchyRequestError: DOM Exception 3.
